// src/app/enums/names/MapName.ts
export enum MapName {
  austriaStates = 'Austria States (9)',
  austriaRegions = 'Austria Regions (34)',
  viennasDistricts = "Vienna's Districts (23)",
  viennasSubDistricts = "Vienna's Sub-Districts (88)",
  austriaCities = 'Austria Cities (Top 100)',
  austriaMountains = 'Austria Mountains',
  austriaLakes = 'Austria Lakes (32)',
  unitedStates = 'United States',
  germanyStates = 'Germany States',
  africaCountries = 'Africa Countries',
  africaCities = 'Africa Cities (Top 25)',
  southAmericaCountries = 'South America Countries',
  southAmericaCities = 'South America Cities (Top 20)',
  europeCountries = "Europe's Countries",
  indiaStates = 'India States',
  indiaCities = 'India Cities (Top 20)',
  brazilStates = 'Brazil States',
  russiaCities = 'Russia Cities (Top 25)',
  biggestCountries = '50 Biggest Countries (by area)',
  topographySouthAmerica = 'Topography South America',
  topographyAsia = 'Topography Asia',
  topographyAfrica = 'Topography Africa'
}